Move physics related actions to another class
